Moved manifests to separate files
