+++ src/game/graphics_handler/manifests/entity_manifest.js
@@ -1,0 +1,45 @@
+import 
+{
+    createBackground,
+    createGenericEntity,
+    createGridObject,
+    createGrid,
+    createTextBox,
+    createPlayer,
+    createCollectible
+} from "../entities/entity_factories.js";
+
+export const ENTITIES = [
+    {
+        typeName: "generic",
+        factoryFunction: createGenericEntity
+    },
+    {
+        typeName: "grid",
+        factoryFunction: createGrid
+    },
+    {
+        typeName: "background",
+        factoryFunction: createBackground
+    },
+    {
+        typeName: "pawn",
+        factoryFunction: createGridObject
+    },
+    {
+        typeName: "textbox",
+        factoryFunction: createTextBox
+    },
+    {
+        typeName: "player",
+        factoryFunction: createPlayer
+    },
+    {
+        typeName: "obstacle",
+        factoryFunction: createGridObject
+    },
+    {
+        typeName: "collectible",
+        factoryFunction: createCollectible
+    }
+]
