+++ src/game/graphics_handler/pixi_renderer.js
@@ -1,5 +1,5 @@
 import * as PIXI from "https://cdnjs.cloudflare.com/ajax/libs/pixi.js/8.1.5/pixi.mjs";
-import { ASSETS } from "./graphics_manifest.js";
+import { ASSETS } from "./manifests/asset_manifest.js";
 
 
 export class PixiRenderer {
